// saevis/src/app/api/model/event.ts
/**
 * Saevis
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: 1.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */import { ModelDate } from './modelDate';
import { User } from './user';


export interface Event { 
    id?: number;
    createdBy?: User;
    updatedAt?: ModelDate;
    createdAt?: ModelDate;
    title: string;
    text: string;
    startTime: Date;
    endTime: Date;
    participants?: Array<User>;
}